+++ samples/repo/src/main/kotlin/com/github/ajalt/clikt/samples/repo/main.kt
@@ -36,7 +36,7 @@ class Repo : CliktCommand(
         for ((k, v) in config) {
             repo.config[k] = v
         }
-        context.obj = repo
+        currentContext.obj = repo
     }
 }